test: migrate SigningOfficialsList to vitest
